requirements.txt
ultralytics 	8.2.75
opencv-python	4.10.0
Pillow		10.4.0

